add menu saiba mais and comissao julgadora
